part2 almost done (check for idle)

+++ labs/lab4/verilog/simulate.sh
@@ -1,5 +1,7 @@
 #!/bin/bash
-rm testbench
-iverilog -s DMATestBench -o testbench DMA_tb.v ramDmaCi.v DMAController.v dualPortSSRAM.v
+if [ -f "testbench" ]; then
+    rm testbench
+fi
+iverilog -s DMATestBench -o testbench BusTransaction_tb.v ramDmaCi.v DMAController.v dualPortSSRAM.v
 ./testbench
 #gtkwave dma_tb.vcd &
